Check for increase in ruff violations and add unit test
Include a script that runs ruff with a lot of rules enabled.
This currently creates a whole bunch of violations in pysweepme.
The script will compare the output to a reference run, which uses
the current code from master (and the rules from the current branch)
and raises an error if for any file-rule combination the amount of
violation increases. That way, we do not need to fix thousands of
violations in one go, and at the same time we ensure that we do
not incorporate more and more violations if we would not check those
rules.

Adding pytest for running unit tests and also include one test
that checks if the pysweepme import does not initialize the
FolderManager yet. This would lead to issues in the multi-instance
feature of SweepMe! and might not be spottet right away.

The tox.ini has the ruffregression in the first place, so that
when running two processes in parallel, it starts rightaway (has the
longest runtime).