Create workflow to build and publish
Add a github workflow that will build and publish pysweepme.
After creating the wheel, pysweepme can be published to PyPI
or TestPyPI. Publishing to PyPI is only allowed from the main
branch. In case of the main branch / upload to PyPI, a tag is
added to the respective commit in the repository.

We do not add tags for TestPyPI or development branches, because
it is not strictly necessary and would just look bad in the
repository after rebases. In particular if multiple rebases
and publications are done alternating.

Credentials for publishing to (Test)PyPI are securely stored
as github Environment Secrets.