Fix and ignore linting problems
Fix issues revealed by ruff for easy-to-fix rules to create a
baseline that can be used from now on and should not fail.

Checking for open TCP/IP ports using the psutil.net_connections()
will not return IP addresses in case of Unix sockets. This case
must to be caught, ruff violations are ignored here.