Include reach function
Reach was previously excluded because the Temperature module
had to identify if a device class defines reach and would also
detect the reach defined in EmptyDevice. Now, the Temperature
module has a more sophisticated check that is able to distinguish
between reach defined in EmptyDevice or in the device class.