Use PortManager to obtain ports in get_driver()
Previously, get_driver was obtaining a (new) port from the
low-level Ports module every time it was called. This led to errors
when the port was already opened, also when combining the same
driver in a SweepMe! module and a CustomFunction Script using
get_driver(). Now, the PortManager is used consistently.